Deduplicate inline `fn sm_uv_flag` to `src/ipred_prepare.rs`.
